requirements.txt
flask
ultralytics
opencv-python
Pillow
torch
torchvision
numpy
diffusers
transformers
accelerate
